Cleanup of reggae-register

+++ templates/reggae-register.sh
@@ -1,12 +1,8 @@
 #!/bin/sh
 
-export INET=$1
-export ACTION=$2
-export IP=$3
-export NAME=`/bin/echo $4 | /usr/bin/cut -f 1 -d '.'`
-export DOMAIN=$5
-export SOCKET="/var/run/reggae/reggae.sock"
-export ZONE_FILE="/usr/local/etc/nsd/zones/master/${DOMAIN}"
+
+SOCKET=${SOCKET:="/var/run/reggae/reggae.sock"}
+ZONE_ROOT=${ZONE_ROOT:="/usr/local/etc/nsd/zones/master"}
 
 
 expand_address() {
@@ -57,29 +53,44 @@ reverse_address() {
 
 
 get_zone() {
-  zone=""
-  address=`expand_address "${1}"`
-  rev_address=`reverse_address ${address}`
-  for char in $(echo ${rev_address} | grep -o .); do
-    zone="${zone}${char}."
-  done
-  echo ${zone}ip6.arpa | cut -b 33-
+  INET="${1}"
+  IP="${2}"
+  if [ "${INET}" = "ipv4" ]; then
+    echo ${IP} | /usr/bin/awk -F '.' '{print $3 "." $2 "." $1 ".in-addr.arpa"}'
+  elif [ "${INET}" = "ipv6" ]; then
+    zone=""
+    address=`expand_address "${IP}"`
+    rev_address=`reverse_address ${address}`
+    for char in $(echo ${rev_address} | grep -o .); do
+      zone="${zone}${char}."
+    done
+    echo ${zone}ip6.arpa | cut -b 33-
+  fi
 }
 
 
 get_ptr() {
-  ptr=""
-  address=`expand_address "${1}"`
-  rev_address=`reverse_address ${address}`
-  for char in $(echo ${rev_address} | grep -o .); do
-    ptr="${ptr}${char}."
-  done
-  echo ${ptr} | cut -b 1-31
+  INET="${1}"
+  IP="${2}"
+  if [ "${INET}" = "ipv4" ]; then
+    echo "${IP}" | /usr/bin/awk -F '.' '{print $4}'
+  elif [ "${INET}" = "ipv6" ]; then
+    ptr=""
+    address=`expand_address "${IP}"`
+    rev_address=`reverse_address ${address}`
+    for char in $(echo ${rev_address} | grep -o .); do
+      ptr="${ptr}${char}."
+    done
+    echo ${ptr} | cut -b 1-31
+  fi
 }
 
 
 create_zone() {
-  cat <<EOF >"${ZONE_FILE}"
+  DOMAIN="${1}"
+  ZONE_FILE="${ZONE_ROOT}/${DOMAIN}"
+  if [ ! -e "${ZONE_FILE}" ]; then
+    cat <<EOF >"${ZONE_FILE}"
 ${DOMAIN}. SOA ${DOMAIN}. hostmaster.DOMAIN. (
                   1998092901  ; Serial number
                   60          ; Refresh
@@ -91,17 +102,26 @@ ${DOMAIN}.            NS      ${DOMAIN}.
 \$ORIGIN ${DOMAIN}.
 EOF
 
-  cat <<EOF >>"/usr/local/etc/nsd/nsd.conf"
+    cat <<EOF >>"/usr/local/etc/nsd/nsd.conf"
 
 zone:
 	name: "${DOMAIN}"
 	zonefile: "master/${DOMAIN}"
 EOF
+
+    /usr/local/bin/sudo /usr/local/sbin/nsd-control reconfig
+  fi
 }
 
 
 create_reverse_zone() {
-  cat <<EOF >"${REVERSE_ZONE_FILE}"
+  INET="${1}"
+  IP="${2}"
+  DOMAIN="${3}"
+  REVERSE_ZONE=`get_zone ${INET} ${IP}`
+  REVERSE_ZONE_FILE="${ZONE_ROOT}/${REVERSE_ZONE}"
+  if [ ! -e "${REVERSE_ZONE_FILE}" ]; then
+    cat <<EOF >"${REVERSE_ZONE_FILE}"
 ${REVERSE_ZONE}.	IN SOA	${DOMAIN}. network.${DOMAIN}. (
 				46         ; serial
 				86400      ; refresh (1 day)
@@ -113,21 +133,30 @@ ${REVERSE_ZONE}.	IN SOA	${DOMAIN}. network.${DOMAIN}. (
 \$ORIGIN ${REVERSE_ZONE}.
 EOF
 
-  cat <<EOF >>"/usr/local/etc/nsd/nsd.conf"
+    cat <<EOF >>"/usr/local/etc/nsd/nsd.conf"
 
 zone:
 	name: "${REVERSE_ZONE}"
 	zonefile: "master/${REVERSE_ZONE}"
 EOF
+
+    /usr/local/bin/sudo /usr/local/sbin/nsd-control reconfig
+  fi
 }
 
 
 cleanup() {
+  INET="${1}"
+  NAME="${2}"
+  IP="${3}"
+  DOMAIN="${4}"
   if [ "${INET}" = "ipv4" ]; then
     export RECORD="A"
   elif [ "${INET}" = "ipv6" ]; then
     export RECORD="AAAA"
   fi
+  IP_REVERSE=`get_ptr ${INET} ${IP}`
+  REVERSE_ZONE=`get_zone ${INET} ${IP}`
   /usr/bin/sed -i "" "/^.* *${RECORD} *${IP}$/d" "${ZONE_FILE}"
   /usr/bin/sed -i "" "/^${NAME} *${RECORD} /d" "${ZONE_FILE}"
   /usr/bin/sed -i "" "/^${IP_REVERSE} *PTR *.*/d" "${REVERSE_ZONE_FILE}"
@@ -136,66 +165,71 @@ cleanup() {
 
 
 alter_host() {
+  INET="${1}"
+  ACTION="${2}"
+  NAME="${3}"
+  IP="${4}"
+  DOMAIN="${5}"
+  IP_REVERSE=`get_ptr ${INET} ${IP}`
+  REVERSE_ZONE=`get_zone ${INET} ${IP}`
   if [ "${INET}" = "ipv4" ]; then
     if [ "${ACTION}" = "add" ]; then
       if [ ! -z "${NAME}" ]; then
-        /bin/echo "${NAME}    A   ${IP}" >>"${ZONE_FILE}"
+        echo "${NAME}    A   ${IP}" >>"${ZONE_FILE}"
         if [ "${NAME}" = "@" ]; then
-          /bin/echo "${IP_REVERSE}    PTR   ${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
+          echo "${IP_REVERSE}    PTR   ${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
         else
-          /bin/echo "${IP_REVERSE}    PTR   ${NAME}.${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
+          echo "${IP_REVERSE}    PTR   ${NAME}.${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
         fi
       fi
-      /bin/echo "register ipv4 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
+      echo "register ipv4 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
+  IP_REVERSE="${5}"
     elif [ "${ACTION}" = "delete" ]; then
-      /bin/echo "unregister ipv4 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
+      echo "unregister ipv4 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
     fi
   elif [ "${INET}" = "ipv6" ]; then
     if [ "${ACTION}" = "add" ]; then
       if [ ! -z "${NAME}" ]; then
-        /bin/echo "${NAME}    AAAA   ${IP}" >>"${ZONE_FILE}"
+        echo "${NAME}    AAAA   ${IP}" >>"${ZONE_FILE}"
         if [ "${NAME}" = "@" ]; then
-          /bin/echo "${IP_REVERSE}    PTR   ${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
+          echo "${IP_REVERSE}    PTR   ${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
         else
-          /bin/echo "${IP_REVERSE}    PTR   ${NAME}.${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
+          echo "${IP_REVERSE}    PTR   ${NAME}.${DOMAIN}." >>"${REVERSE_ZONE_FILE}"
         fi
       fi
-      /bin/echo "register ipv6 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
+      echo "register ipv6 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
     elif [ "${ACTION}" = "delete" ]; then
-      /bin/echo "unregister ipv6 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
+      echo "unregister ipv6 ${IP} ${NAME}${DOMAIN}" | /usr/bin/nc -U "${SOCKET}" -w 0
     fi
   fi
+  /usr/local/bin/sudo /usr/local/sbin/nsd-control reload ${DOMAIN}
+  /usr/local/bin/sudo /usr/local/sbin/nsd-control reload ${REVERSE_ZONE}
 }
 
 
-if [ "${INET}" = "ipv4" ]; then
-  export REVERSE_ZONE=`/bin/echo ${IP} | /usr/bin/awk -F '.' '{print $3 "." $2 "." $1 ".in-addr.arpa"}'`
-elif [ "${INET}" = "ipv6" ]; then
-  export REVERSE_ZONE=`get_zone ${IP}`
+main() {
+  INET="${1}"
+  ACTION="${2}"
+  IP="${3}"
+  NAME="${4}"
+  DOMAIN="${5}"
+
+  create_zone "${DOMAIN}"
+  create_reverse_zone "${INET}" "${IP}" "${DOMAIN}"
+  cleanup "${INET}" "${NAME}" "${IP}" "${DOMAIN}"
+  alter_host "${INET}" "${ACTION}" "${NAME}" "${IP}" "${DOMAIN}"
+}
+
+
+if [ "$#" = "5" ]; then
+  INET="${1}"
+  ACTION="${2}"
+  IP="${3}"
+  NAME=`echo ${4} | /usr/bin/cut -f 1 -d '.'`
+  DOMAIN="${5}"
+
+  main "${INET}" "${ACTION}" "${IP}" "${NAME}" "${DOMAIN}"
+elif [ "$#" != "0" ]; then
+  echo "Usage: $0 <ipv4/ipv6> <add/delete> <ip> <name> <domain>" >&2
+  exit 1
 fi
-export REVERSE_ZONE_FILE="/usr/local/etc/nsd/zones/master/${REVERSE_ZONE}"
-
-
-if [ "${INET}" = "ipv4" ]; then
-  export IP_REVERSE=`/bin/echo "${IP}" | /usr/bin/awk -F '.' '{print $4}'`
-elif [ "${INET}" = "ipv6" ]; then
-  export IP_REVERSE=`get_ptr ${IP}`
-fi
-
-
-if [ ! -e "${ZONE_FILE}" ]; then
-  create_zone
-  /usr/local/bin/sudo /usr/sbin/service nsd restart
-fi
-
-if [ ! -e "${REVERSE_ZONE_FILE}" ]; then
-  create_reverse_zone
-  /usr/local/bin/sudo /usr/sbin/service nsd restart
-fi
-
-
-cleanup
-alter_host
-
-/usr/local/bin/sudo /usr/local/sbin/nsd-control reload ${DOMAIN}
-/usr/local/bin/sudo /usr/local/sbin/nsd-control reload ${REVERSE_ZONE}
